teleport, sound and players moving

+++ Monopoly/GradleProject/Monopoly/app/src/main/java/com/rob/monopoly/Interfaces/Improvable.java
@@ -8,7 +8,7 @@ public interface Improvable extends Ownable {
 
     boolean buildHouse();
     boolean demolishHouse();
-    double getHousePrice();
+    int getHousePrice();
     double getNumHouses();
     String getColourGroup();
     boolean ownsColourGroup();
